vue-web/Interface/requirements.txt
PyExecJS==1.5.1
requests==2.27.1
